Fix runtime with python
