+++ csrcs/fastdeploy/core/fd_type.cc
@@ -17,7 +17,7 @@
 
 namespace fastdeploy {
 
-int FDDataTypeSize(FDDataType data_type) {
+int FDDataTypeSize(const FDDataType& data_type) {
   FDASSERT(data_type != FDDataType::FP16, "Float16 is not supported.");
   if (data_type == FDDataType::BOOL) {
     return sizeof(bool);
@@ -34,89 +34,63 @@ int FDDataTypeSize(FDDataType data_type) {
   } else if (data_type == FDDataType::UINT8) {
     return sizeof(uint8_t);
   } else {
-    FDASSERT(false, "Unexpected data type: " + FDDataTypeStr(data_type));
+    FDASSERT(false, "Unexpected data type: " + Str(data_type));
   }
   return -1;
 }
 
-std::string FDDataTypeStr(FDDataType data_type) {
-  FDASSERT(data_type != FDDataType::FP16, "Float16 is not supported.");
-  if (data_type == FDDataType::BOOL) {
-    return "bool";
-  } else if (data_type == FDDataType::INT16) {
-    return "int16";
-  } else if (data_type == FDDataType::INT32) {
-    return "int32";
-  } else if (data_type == FDDataType::INT64) {
-    return "int64";
-  } else if (data_type == FDDataType::FP16) {
-    return "float16";
-  } else if (data_type == FDDataType::FP32) {
-    return "float32";
-  } else if (data_type == FDDataType::FP64) {
-    return "float64";
-  } else if (data_type == FDDataType::UINT8) {
-    return "uint8";
-  } else if (data_type == FDDataType::INT8) {
-    return "int8";
-  } else {
-    FDASSERT(false, "Unexpected data type: " + FDDataTypeStr(data_type));
-  }
-  return "UNKNOWN!";
-}
-
-std::string Str(Device& d) {
+std::string Str(const Device& d) {
   std::string out;
   switch (d) {
-  case Device::DEFAULT:
-    out = "Device::DEFAULT";
-    break;
-  case Device::CPU:
-    out = "Device::CPU";
-    break;
-  case Device::GPU:
-    out = "Device::GPU";
-    break;
-  default:
-    out = "Device::UNKOWN";
+    case Device::DEFAULT:
+      out = "Device::DEFAULT";
+      break;
+    case Device::CPU:
+      out = "Device::CPU";
+      break;
+    case Device::GPU:
+      out = "Device::GPU";
+      break;
+    default:
+      out = "Device::UNKOWN";
   }
   return out;
 }
 
-std::string Str(FDDataType& fdt) {
+std::string Str(const FDDataType& fdt) {
   std::string out;
   switch (fdt) {
-  case FDDataType::BOOL:
-    out = "FDDataType::BOOL";
-    break;
-  case FDDataType::INT16:
-    out = "FDDataType::INT16";
-    break;
-  case FDDataType::INT32:
-    out = "FDDataType::INT32";
-    break;
-  case FDDataType::INT64:
-    out = "FDDataType::INT64";
-    break;
-  case FDDataType::FP32:
-    out = "FDDataType::FP32";
-    break;
-  case FDDataType::FP64:
-    out = "FDDataType::FP64";
-    break;
-  case FDDataType::FP16:
-    out = "FDDataType::FP16";
-    break;
-  case FDDataType::UINT8:
-    out = "FDDataType::UINT8";
-    break;
-  case FDDataType::INT8:
-    out = "FDDataType::INT8";
-    break;
-  default:
-    out = "FDDataType::UNKNOWN";
+    case FDDataType::BOOL:
+      out = "FDDataType::BOOL";
+      break;
+    case FDDataType::INT16:
+      out = "FDDataType::INT16";
+      break;
+    case FDDataType::INT32:
+      out = "FDDataType::INT32";
+      break;
+    case FDDataType::INT64:
+      out = "FDDataType::INT64";
+      break;
+    case FDDataType::FP32:
+      out = "FDDataType::FP32";
+      break;
+    case FDDataType::FP64:
+      out = "FDDataType::FP64";
+      break;
+    case FDDataType::FP16:
+      out = "FDDataType::FP16";
+      break;
+    case FDDataType::UINT8:
+      out = "FDDataType::UINT8";
+      break;
+    case FDDataType::INT8:
+      out = "FDDataType::INT8";
+      break;
+    default:
+      out = "FDDataType::UNKNOWN";
   }
   return out;
 }
 
-} // namespace fastdeploy
+}  // namespace fastdeploy
