+++ csrcs/fastdeploy/vision/ppdet/ppyoloe.h
@@ -42,6 +42,10 @@ class FASTDEPLOY_DECL PPYOLOE : public FastDeployModel {
   int64_t nms_top_k = 10000;
   bool normalized = true;
   bool has_nms_ = false;
+
+  // This function will used to check if this model contains multiclass_nms
+  // and get parameters from the operator
+  void GetNmsInfo();
 };
 }  // namespace ppdet
 }  // namespace vision
